# api/openapi.yaml
openapi: 3.1.0
info:
  title: ConsorcioAbierto API
  version: 0.1.0
  description: |
    API del SaaS multi-tenant de administración de consorcios.
    Contrato de Fase 0. Convenciones: JSON snake_case, fechas ISO 8601,
    importes como enteros `*_cents` + `currency`, IDs UUID.
    Errores RFC 9457 (application/problem+json).
servers:
  - url: /api/v1
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ca_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: Error RFC 9457.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ErrorField:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
...
    User:
      type: object
      required: [id, email, name, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        status: { type: string }
        mfa_enabled: { type: boolean }
    Membership:
      type: object
      required: [id, tenant_id, tenant_name, roles]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        tenant_name: { type: string }
        tenant_status: { type: string }
        roles:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
    AuthResponse:
      type: object
      required: [user, memberships]
      properties:
        user:
          $ref: "#/components/schemas/User"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
...
    MfaSetup:
      type: object
      required: [secret, otpauth_url]
      properties:
        secret: { type: string, description: Secret TOTP, se muestra una sola vez }
        otpauth_url: { type: string, description: URL otpauth:// para escanear con el autenticador }
...
    Me:
      type: object
      required: [user, membership, permissions]
      properties:
        user:
          $ref: "#/components/schemas/User"
        membership:
          $ref: "#/components/schemas/Membership"
        permissions:
          type: array
          items:
            type: string
...
paths:
...
  /auth/select-tenant:
    post:
      operationId: selectTenant
      tags: [auth]
      security:
        - cookieAuth: []
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [membership_id]
              properties:
                membership_id: { type: string, format: uuid }
      responses:
        "200":
          description: Emite access token contextual a la membresía elegida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "403": { $ref: "#/components/responses/Problem" }
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      responses:
        "200":
          description: Rota refresh y emite nuevo access token.
        "401": { $ref: "#/components/responses/Problem" }
...
  /auth/mfa/setup:
    post:
      operationId: mfaSetup
      tags: [auth]
      security:
        - cookieAuth: []
          bearerAuth: []
      responses:
        "200":
          description: Genera un secret TOTP y devuelve la URL otpauth para el QR.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaSetup"
        "401": { $ref: "#/components/responses/Problem" }
  /auth/mfa/confirm:
    post:
      operationId: mfaConfirm
      tags: [auth]
      security:
        - cookieAuth: []
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, description: Código TOTP de 6 dígitos }
      responses:
        "204": { description: MFA habilitado. }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /auth/mfa/verify:
    post:
      operationId: mfaVerify
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token: { type: string, description: Token de un solo propósito emitido por /auth/login }
                code: { type: string, description: Código TOTP de 6 dígitos }
      responses:
        "200":
          description: Segundo factor válido; emite refresh + access.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/mfa/disable:
    post:
      operationId: mfaDisable
      tags: [auth]
      security:
        - cookieAuth: []
          bearerAuth: []
      responses:
        "204": { description: MFA deshabilitado. }
        "401": { $ref: "#/components/responses/Problem" }
...